room_temperature_sensor/requirements.txt
paho-mqtt==1.6.1
pyyaml==6.0
pytest==7.3.1
